front: extract syn-string utilities into front::syn_metadata
Move the two leaf helpers `transparent_result_ok_type` (used by
`jit_codewriter::call`'s `?`-desugar transparency check) and
`nolength_from_array_type_id` (used by the same caller for
length-prefix shape detection) plus the helper
`first_top_level_generic_arg` they depend on, into a new
`front/syn_metadata.rs` module.

Both are pure string-manipulation utilities with no
`GraphBuildContext` dependency, so they belong outside the
syn-AST graph builder.  `front::ast` retains thin `pub(crate) use`
re-exports for in-crate callers; the function bodies move out.

This is the start of carving small chunks off `front::ast`
ahead of Step 6.F (full deletion).  Larger metadata collectors
(`collect_program_metadata_pub`, `collect_jit_hints_with_sig`,
`collect_struct_origins`) stay in `ast.rs` for now because they
share internal helpers with the graph builder.
