+++ majit/majit-translate/src/front/ast.rs
@@ -10824,40 +10824,7 @@ fn split_tuple_type_elements(type_str: &str) -> Option<Vec<String>> {
 /// retains a length header at offset 0; a `*const GcArray<T>` is a
 /// pointer to that header block, not to items[0], so its descr must
 /// carry `lendescr` per `descr.py:359-362`.
-pub(crate) fn nolength_from_array_type_id(array_type_id: Option<&str>) -> bool {
-    let Some(s) = array_type_id else {
-        return false;
-    };
-    let mut inner = s.trim();
-    loop {
-        let stripped = inner
-            .strip_prefix("*const ")
-            .or_else(|| inner.strip_prefix("*mut "))
-            .or_else(|| inner.strip_prefix("&mut "))
-            .or_else(|| inner.strip_prefix('&'));
-        match stripped {
-            Some(rest) => inner = rest.trim_start(),
-            None => break,
-        }
-    }
-    // `[T]` / `[T; N]` are unambiguous headerless item runs.
-    if inner.starts_with('[') && inner.ends_with(']') {
-        return true;
-    }
-    // Length-prefixed wrappers carry `<` (generic) or `(` (paren-style
-    // lltype spelling such as `Ptr(GcArray(...))`).  Keep the PyPy
-    // default `False` for those — a pointer to a wrapper still
-    // dereferences a length header.
-    if inner.contains('<') || inner.contains('(') {
-        return false;
-    }
-    // Bare identifier pointee (`*const i64`, `*const Point`) means the
-    // pointer addresses items[0] of a primitive / struct item type.
-    // A bare identifier with NO pointer prefix is a value-type binding
-    // (e.g. an `array_type_id` directly naming a struct that contains
-    // an embedded array); preserve the PyPy default `False` for that.
-    s.trim() != inner
-}
+pub(crate) use crate::front::syn_metadata::nolength_from_array_type_id;
 
 fn bind_ident_type(ident: &syn::Ident, type_str: &str, ctx: &mut GraphBuildContext<'_>) {
     let name = ident.to_string();
@@ -11569,23 +11536,7 @@ fn type_string_to_value_type(type_str: &str) -> ValueType {
     }
 }
 
-pub(crate) fn transparent_result_ok_type(type_str: &str) -> Option<&str> {
-    let trimmed = type_str.trim();
-    for prefix in ["Result<", "std::result::Result<", "core::result::Result<"] {
-        let Some(inner) = trimmed
-            .strip_prefix(prefix)
-            .and_then(|rest| rest.strip_suffix('>'))
-        else {
-            continue;
-        };
-        let ok_type = first_top_level_generic_arg(inner).map(str::trim)?;
-        if ok_type == "()" {
-            return None;
-        }
-        return Some(ok_type);
-    }
-    None
-}
+pub(crate) use crate::front::syn_metadata::transparent_result_ok_type;
 
 fn transparent_result_err_type(type_str: &str) -> Option<&str> {
     let trimmed = type_str.trim();
