+++ majit/majit-translate/src/front/syn_metadata.rs
@@ -1,0 +1,98 @@
+//! Small syn-AST utility helpers extracted from `front::ast`.
+//!
+//! These are pure string-manipulation / attribute-walking helpers
+//! that do not depend on `GraphBuildContext` or any other piece of
+//! the AST graph builder.  Carving them out is a step toward
+//! retiring `front::ast` (issue #97 Step 6.F): consumers outside the
+//! graph builder (`jit_codewriter::call`, hybrid passes in `lib.rs`)
+//! can depend on this module without pulling in 15k LOC of legacy
+//! AST lowering.
+
+/// Detect the canonical `Result<T, …>` wrapper and project the inner
+/// `T`.  Returns `None` for non-`Result` shapes, for `Result<(), …>`
+/// (no transparent type to project), and for malformed inputs.
+///
+/// `front::ast::transparent_result_ok_type` was the original home;
+/// the only consumers (`jit_codewriter::call`) live outside the AST
+/// graph builder so the helper moves with them.
+pub fn transparent_result_ok_type(type_str: &str) -> Option<&str> {
+    let trimmed = type_str.trim();
+    for prefix in ["Result<", "std::result::Result<", "core::result::Result<"] {
+        let Some(inner) = trimmed
+            .strip_prefix(prefix)
+            .and_then(|rest| rest.strip_suffix('>'))
+        else {
+            continue;
+        };
+        let ok_type = first_top_level_generic_arg(inner).map(str::trim)?;
+        if ok_type == "()" {
+            return None;
+        }
+        return Some(ok_type);
+    }
+    None
+}
+
+/// Return the first comma-delimited top-level generic argument in
+/// `args` (`"A, B<C, D>, E"` → `"A"`).  Tracks bracket depth so a
+/// nested generic boundary does not confuse the split.
+///
+/// Lives here because [`transparent_result_ok_type`] needs it; the
+/// AST graph builder retains its own copy under
+/// `front::ast::first_top_level_generic_arg` for its private callers
+/// (`transparent_result_err_type`, `transparent_option_inner_type`).
+pub fn first_top_level_generic_arg(args: &str) -> Option<&str> {
+    let mut depth = 0usize;
+    for (idx, ch) in args.char_indices() {
+        match ch {
+            '<' | '(' | '[' => depth += 1,
+            '>' | ')' | ']' => depth = depth.saturating_sub(1),
+            ',' if depth == 0 => return Some(&args[..idx]),
+            _ => {}
+        }
+    }
+    if args.is_empty() { None } else { Some(args) }
+}
+
+/// Decide whether a registered `array_type_id` describes a
+/// headerless item-run pointee or a length-prefixed wrapper.  Bare
+/// pointers to identifier types address `items[0]` (no length word);
+/// `Vec<T>` / `GcArray<T>` / `Ptr(GcArray(T))` shapes carry a length
+/// header at offset 0 and therefore keep the PyPy default `False`.
+///
+/// Originally `front::ast::nolength_from_array_type_id`; only
+/// `jit_codewriter::call` consumes it, and it carries no AST-builder
+/// dependency.
+pub fn nolength_from_array_type_id(array_type_id: Option<&str>) -> bool {
+    let Some(s) = array_type_id else {
+        return false;
+    };
+    let mut inner = s.trim();
+    loop {
+        let stripped = inner
+            .strip_prefix("*const ")
+            .or_else(|| inner.strip_prefix("*mut "))
+            .or_else(|| inner.strip_prefix("&mut "))
+            .or_else(|| inner.strip_prefix('&'));
+        match stripped {
+            Some(rest) => inner = rest.trim_start(),
+            None => break,
+        }
+    }
+    if inner.starts_with('[') && inner.ends_with(']') {
+        return true;
+    }
+    // Length-prefixed wrappers carry `<` (generic) or `(` (paren-style
+    // lltype spelling such as `Ptr(GcArray(...))`).  Keep the PyPy
+    // default `False` for those — a pointer to a wrapper still
+    // dereferences a length header.
+    if inner.contains('<') || inner.contains('(') {
+        return false;
+    }
+    // Bare identifier pointee (`*const i64`, `*const Point`) means the
+    // pointer addresses items[0] of a primitive / struct item type.
+    // A bare identifier with NO pointer prefix is a value-type binding
+    // (e.g. an `array_type_id` directly naming a struct that contains
+    // an embedded array); preserve the PyPy default `False` for that.
+    s.trim() != inner
+}
